Organized the code adding controllers for readability & debugging purposes
